Excluindo imports desnecessários nas views e ajustando alguns forms
