+++ capacita/templates/registration/login.html
