+++ capacitaApp/forms.py
@@ -4,24 +4,6 @@ from django import forms
 from django.contrib.auth.models import User, Group
 from .models import *
 
-class OrgaoForm(forms.ModelForm):
-
-    class Meta:
-        model = Orgao
-        fields = ('nome', )
-        labels = {
-            'nome' : 'Nome'
-        }
-
-class ObjetivoTreinamentoForm(forms.ModelForm):
-
-    class Meta:
-        model = Objetivo_Treinamento
-        fields = ('nome', )
-        labels = {
-            'nome' : 'Nome'
-        }
-
 class ProfileForm(forms.ModelForm):
 
     class Meta:
